VIX-3871 Fix leftover items from last scan
* Redundant type casts that were missed.
* Missed async await issue in setup form.
* Missed null attribute not matching override.

+++ src/Vixen.Modules/Effect/Pattern/PatternBrick.cs
@@ -147,7 +147,7 @@ namespace VixenModules.Effect.Pattern
 			int count = 2;
 			
 			// Need the bar thickness then the bar spacing for each of the brick color bars
-			return (int)(count * brickHeight) + count * mortorHeight;			
+			return count * brickHeight + count * mortorHeight;			
 		}
 
 		/// <summary>
